requirements.txt
asgiref==3.6.0
crispy-bootstrap4==2022.1
Django==4.2
django-ckeditor==6.5.1
django-crispy-forms==2.0
django-js-asset==2.0.0
psycopg2==2.9.6
sqlparse==0.4.4
tzdata==2023.3
